main menu nav update
